+++ packages/superflare-remix-dev/.gitignore
@@ -1,0 +1,1 @@
+dist
