+++ packages/superflare-remix-dev/README.md
@@ -1,0 +1,9 @@
+# @superflare/remix-dev
+
+This package provides `superflareDevProxyVitePlugin`, a drop-in replacement for remix’s [`cloudflareDevProxyVitePlugin`](https://remix.run/docs/en/main/guides/vite#cloudflare-proxy) that adds support for the superflare feature set to Vite’s dev server.
+
+[Read the docs](https://superflare.dev), or get started now:
+
+```bash
+npx superflare@latest new
+```
