+++ packages/superflare-remix-dev/cloudflare.env.d.ts
@@ -1,0 +1,1 @@
+/// <reference types="@cloudflare/workers-types" />
