+++ packages/superflare-remix-dev/tsup.config.ts
@@ -1,0 +1,19 @@
+import { defineConfig } from "tsup";
+
+const nodejsCompatPlugin: any = {
+  name: "nodejs_compat Plugin",
+  setup(pluginBuild: any) {
+    pluginBuild.onResolve({ filter: /node:.*/ }, (opts: any) => {
+      return { external: true };
+    });
+  },
+};
+
+export default defineConfig([
+  {
+    format: ["esm", "cjs"],
+    esbuildPlugins: [nodejsCompatPlugin],
+    entry: ["index.ts"],
+    dts: true,
+  },
+]);
